Feat: 알림 DB 관계 변경으로 인한 수정 및 FCM 토큰 관련 수정
- FCM 토큰 저장 시 토큰 중복 체크
- 알림 숫자 조회 시 읽지 않은 알림만 조회되도록 수정

+++ src/main/java/com/notificationserver/adaptor/in/web/vo/NotificationLogCountVo.java
@@ -12,7 +12,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class NotificationLogCountVo {
 
-	private Integer notificationLogCount;
+	private long notificationLogCount;
 
 	public static NotificationLogCountVo getNotificationLogInDto(NotificationLogCountInDto dto) {
 		return NotificationLogCountVo.builder()
